More work on swri_image_util; it's good enough for now
