+++ swri_image_util/src/nodes/normalize_response_node.cpp
