+++ swri_image_util/src/nodes/replace_colors_node.cpp
